requirements.txt
fastapi==0.74.1
pydantic==1.9.0
PyJWT==2.3.0
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
uvicorn==0.17.5
passlib==1.7.4
python-dotenv==0.21.0
pymysql==1.0.2
requests
geopy